dice-protocols: tcat: extension: code refactoring to obsolete FormatEntryData
As well as router entries, byte data for stream entry is used instead of
strutured data. It's inconvenient for runtime to recompile the data to
structured data every time.

This commit obsoletes the usage of FormatEntryData in runtime crate.
